requirements.txt
aiofiles==23.2.1
aiogram==3.13.1
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.10.4
async-timeout==4.0.3
asyncpg==0.29.0
attrs==25.3.0
Babel==2.9.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
Flask==3.1.2
frozenlist==1.7.0
google-auth==2.29.0
google-auth-oauthlib==1.2.2
gspread==6.1.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.27.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
magic-filter==1.0.12
MarkupSafe==3.0.3
multidict==6.6.4
oauth2client==4.1.3
oauthlib==3.3.1
propcache==0.3.2
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.2.5
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
transliterate==1.10.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1
